Replace ST_(M)ShortestPathLength by new implementation.
