+++ src/main/java/org/gdms/gdmstopology/Activator.java
@@ -34,15 +34,13 @@ import org.gdms.gdmstopology.function.ST_BlockIdentity;
 import org.gdms.gdmstopology.centrality.ST_GraphAnalysis;
 import org.gdms.gdmstopology.centrality.ST_StrahlerStreamOrder;
 import org.gdms.gdmstopology.function.ST_ConnectedComponents;
-import org.gdms.gdmstopology.function.ST_Distances;
+import org.gdms.gdmstopology.function.ST_ShortestPathLength;
 import org.gdms.gdmstopology.function.ST_FindReachableEdges;
 import org.gdms.gdmstopology.function.ST_Graph;
 import org.gdms.gdmstopology.function.ST_MFindReachableEdges;
 import org.gdms.gdmstopology.function.ST_MShortestPath;
-import org.gdms.gdmstopology.function.ST_MShortestPathLength;
 import org.gdms.gdmstopology.function.ST_PlanarGraph;
 import org.gdms.gdmstopology.function.ST_ShortestPath;
-import org.gdms.gdmstopology.function.ST_ShortestPathLength;
 import org.gdms.gdmstopology.function.ST_SubGraphStatistics;
 import org.gdms.gdmstopology.function.ST_ToLineNoder;
 import org.gdms.sql.function.Function;
@@ -65,9 +63,7 @@ public class Activator implements BundleActivator {
         reg(new ST_Graph());
         reg(new ST_PlanarGraph());
         reg(new ST_ShortestPath());
-        reg(new ST_ShortestPathLength());
         reg(new ST_ToLineNoder());
-        reg(new ST_MShortestPathLength());
         reg(new ST_FindReachableEdges());
         reg(new ST_MFindReachableEdges());
         reg(new ST_MShortestPath());
@@ -75,7 +71,7 @@ public class Activator implements BundleActivator {
         reg(new ST_ConnectedComponents());
         reg(new ST_GraphAnalysis());
         reg(new ST_StrahlerStreamOrder());
-        reg(new ST_Distances());
+        reg(new ST_ShortestPathLength());
     }
 
     private void reg(Function gdmsFunc) {
